+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+10/6/2020
++ Edit Note functionality
++ Delete Note functionality
++ Styling
+
+10/5/2020
++ Add Note functionality
 
 10/5/2020
 + Initial commit react setup.